Removed mention of REST API in top comment as this is enabled by default in JIRA.
Updated jqlTemplate - if you're delivering a version of a product you
want to know the status of all of the issues associated with that
fixVersion, not just the ones which are Resolved/Closed.

Minor formatting changes in comments (suspicious of Eclipse's
Auto-formatting feature for changing these).

Increased the maximum number of issues to 500.

Added various log messages.